rename auth api and fix some bugs
* add 'auth' category before all auth APIs
* fix unable to removeEdge() when enable auth
* fix delete admin user(expect not allowed)
* check auth element not exist before creating
* throw 400 instead of 404 when delete non-exist element
